Restructuring + more query params-related fixes
