+++ apps/ensadmin/src/app/@actions/example-queries/page.tsx
@@ -1,4 +1,4 @@
-import ExampleQueriesDocumentationLinks from "@/app/example-queries/subcomponents/ExampleQueriesDocumentationLinks";
+import ExampleQueriesDocumentationLinks from "@/app/example-queries/components/ExampleQueriesDocumentationLinks";
 
 //TODO: for the sake of responsiveness this elem was moved into the main panel for smaller widths (<1280px) but that might change if the panel's overall design changes
 export default function ActionsExampleQueries() {
