+++ apps/ensadmin/src/app/example-queries/example-editor/page.tsx
@@ -1,5 +1,7 @@
 import { ExampleQueriesGraphiQLEditor } from "@/components/graphiql-editor";
 import { defaultEnsNodeUrl } from "@/lib/env";
+import ExampleQueriesDocumentationLinks from "@/app/example-queries/components/ExampleQueriesDocumentationLinks";
+import React from "react";
 
 //TODO: this type is defined many times across multiple pages (I assume to avoid exporting) but maybe we can unify it?
 type PageProps = {
@@ -20,10 +22,13 @@ export default async function ExampleQueryEditor({ searchParams }: PageProps) {
   const url = new URL(`/subgraph`, baseUrl).toString();
 
   return (
-    <ExampleQueriesGraphiQLEditor
-      url={url}
-      query={query as string}
-      variables={variables && (variables as string)}
-    />
+      <section className="flex flex-col flex-1">
+        <ExampleQueriesDocumentationLinks styles="w-fit flex xl:hidden bg-gray-100 p-4 rounded-lg mt-4 mx-4" />
+        <ExampleQueriesGraphiQLEditor
+            url={url}
+            query={query as string}
+            variables={variables && (variables as string)}
+        />
+      </section>
   );
 }
